+++ src/textdisplay.cc
@@ -84,7 +84,7 @@ ostream &operator<<(ostream &out, const TextDisplay &td){
 	ostringstream convert;
 	
 	string race = Player::charToRace(player->getSymbol());
-	int gold = player->getGold();
+	double gold = player->getGold();
 	int hp = player->getHp();
 	if (hp < 0) hp=0; //as to not display negative hp
 	int atk = player->getAtk();
